pyproject.toml
[project]
name = "engenharia-softwareii-2025-2-t04-crawl4ai"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.13"
dependencies = [
    "gitpython>=3.1.45",
]
